requirements.txt
flask
azure-identity
azure-storage-blob
python-dotenv
